requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
cycler==0.10.0
idna==2.7
kiwisolver==1.0.1
matplotlib==3.0.0
more-itertools==4.3.0
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
pluggy==0.8.0
py==1.7.0
pyparsing==2.2.2
pytest==3.9.2
python-dateutil==2.7.3
requests==2.20.0
six==1.11.0
urllib3==1.24
